requirements.txt
alabaster==1.0.0
babel==2.17.0
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.1
cryptography==46.0.3
docutils==0.21.2
flake8==7.3.0
idna==3.11
imagesize==1.4.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2

pyodbc==5.3.0
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
RapidFuzz==3.14.3
requests==2.32.5
roman-numerals-py==3.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
